Link RHEL5 SRPM 1.6 release pages
The generic SRPM we currently link doesn't work on RHEL5 and similarly
old systems, since RHEL5 RPM doesn't understand non-MD5 checksums. We
have one that works on RHEL5, though, so link it.

Do this for releases 1.6.2, 1.6.4, and 1.6.5. The SRPM for releases
older than 1.6.2 appear to work on RHEL5, and 1.6.2.1 doesn't have any
RHEL5-specific files, since 1.6.2 and 1.6.2.1 are identical on RHEL5.

Change-Id: Ie4bd211c4d0aba8aa0272c38a0d640c20fddb408
Reviewed-on: http://gerrit.openafs.org/10125
Reviewed-by: Ken Dreyer <ktdreyer@ktdreyer.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: Ken Dreyer <ktdreyer@ktdreyer.com>