Rototill roadmap.html
Update the upcoming versions and move those to a more prominent position,
since they are more certain and well-understood.

Greatly trim the wish list of features, since there is not currently
any indication that most of them will happen in the near future.
List rxgk and IPv6 as the biggest/most pressing issues.

Change-Id: I44db34ae30487e081d846ff427f948c2f04dd53b
Reviewed-on: https://gerrit.openafs.org/12417
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Tested-by: Benjamin Kaduk <kaduk@mit.edu>