support: modify sinenomine.net links to work in frame
www.sinenomine.net is configured to not serve content to a frame.
Therefore the links on openafs.org (which uses frames) don't open with a
direct click, but only when opened in a new window or tab.

Add a target attribute so the link will open when clicked directly.
While here, modify the URL to point at the OpenAFS content on the site.

Change-Id: I6365c0afa9e051ce5c9813a7eb449ee6cafa3fcc
Reviewed-on: https://gerrit.openafs.org/13015
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Stephan Wiesand <stephan.wiesand@desy.de>
Tested-by: Stephan Wiesand <stephan.wiesand@desy.de>